refactor: devnet runtime
