+++ runtime/devnet/src/config/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod api;
 pub mod assets;
 mod contracts;
 mod proxy;
